Enhance dep-relations for paasng.platform.accounts, more
