+++ apiserver/paasng/.importlinter
@@ -31,23 +31,15 @@ layers =
     paasng.accessories.bk_components | paasng.accounts.permissions
     paasng.accessories.iam 
     paasng.accounts.models
-    paasng.accounts.oauth | paasng.platform.feature_flags
-    paasng.platform.region
-    paasng.dev_resources.templates
-    paasng.platform.core 
+    paasng.dev_resources.templates | paasng.accounts.oauth | paasng.platform.feature_flags
+    paasng.platform.region | paasng.platform.core 
 ignore_imports =
     # TODO: Remove these rules after the dependency relations are fixed.
     paasng.platform.applications.models -> paasng.publish.**
     paasng.platform.applications.models -> paasng.accessories.iam.**
     paasng.platform.applications.views_ext -> paasng.plat_admin.system.applications
-    paasng.platform.region.permissions -> paasng.utils.api_docs
-    paasng.platform.region.permissions -> paasng.accounts.models
-    paasng.utils.models -> paasng.platform.region.states
-    paasng.utils.models -> paasng.platform.region.models
-    paasng.utils.validators -> paasng.platform.region.models
     paasng.accounts.oauth.* -> paasng.dev_resources.sourcectl.*
     paasng.accounts.oauth.backends -> paasng.platform.applications.models
-    paasng.accessories.iam.permissions.resources.application -> paasng.accounts.permissions.constants
     paasng.dev_resources.sourcectl.connector -> paasng.dev_resources.services.utils
     paasng.extensions.scene_app.initializer -> paasng.extensions.declarative.*
     paasng.accessories.paas_analysis.utils -> paasng.platform.environments.models
