+++ apiserver/paasng/paasng/accessories/iam/constants.py
@@ -134,3 +134,8 @@ APP_MINI_ACTIONS_IN_BK_LOG = {
         ],
     },
 }
+
+
+# 由于权限中心的用户组授权为异步行为，即创建用户组，添加用户，对组授权后需要等待一段时间（10-20秒左右）才能鉴权
+# 因此需要在应用创建后的一定的时间内，对创建者（拥有应用最高权限）的操作进行权限豁免以保证功能可正常使用
+PERM_EXEMPT_TIME_FOR_OWNER_AFTER_CREATE_APP = 5 * 60
